services/selenium-daemon/requirements.txt
nodriver>=0.38
fastapi>=0.110
uvicorn[standard]>=0.29
google-cloud-firestore>=2.16
google-cloud-storage>=2.16
google-auth>=2.29
pyOpenSSL>=24.0
google-generativeai>=0.5
google-cloud-secret-manager>=2.18
Pillow>=10.3
numpy>=1.26
psutil>=5.9
pyyaml>=6.0
aiohttp>=3.9
sendgrid>=6.11
pypdf>=4.0
